引用楼盘结构考核任务处理

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/BasicApplyBatchDetailMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface BasicApplyBatchDetailMapper {
-    int countByExample(BasicApplyBatchDetailExample example);
+    long countByExample(BasicApplyBatchDetailExample example);
 
     int deleteByExample(BasicApplyBatchDetailExample example);
